Move router and subgraph functions to separate files for readability
